refactor: more porting to lua
removed some a few configruation files that seemed to be obsolete e.g.
the yank highlight time which was set to the default.

I'm also going back and forth on waht my convention should be for the
lua module init file. I dont' know if that should only require it's
associated children or if it should also include inherent global/core
settings as well. For example I am trying to decide if
mappings/global.lua should exist or if those should be in init.lua. I
suspect there is a lua convention that I need to read up on.