feat(plugins): telescope plugin frecency
I  find myself regularly re-searching for the same file(s) and this
plugins seems to be a better solution to that. I am cautious that
this may be a workflow shortcoming that can be better managed or very
likely to be part of the dotfile tinkering workflow which is abnormal.