refactor(omnisharp): replace omnisharp through lsp with omnisharp-vim
this refactor isn't without it's issues. While it does provide
a lot more features while editing C# it's done at the expense of some
ugly plugin dancing. namely the completion seems like a hack. In a C#
file compe is disabled and deoplete takes over. This seems to work
fine during testing but we'll see if this causes issues when it's being
used under real world conditions.

there are also some other changes that shouldn't really be part of this.