refactor(dotfiles): add bashrc and cleansup zshrc
On some machines bashrc was being appended to unreliably. Pull it into
versioned files and simply the process.

Cleanup a bunch of comments in the zshrc

I did find an interesting sumneko alias, one of which was wrong. I'm not
sure how that gets added if it was my mistake or it's done during the
build process.